backend/requirements.txt
# Python dependencies for the Altera Labs backend.
# Core dependencies - installed first for faster container builds
flask
flask-cors
python-dotenv
requests
PyMuPDF
pexpect
tqdm
requests_toolbelt
protobuf>=3.20.0
google-genai>=1.30.0
rich>=13.7.1
pandas>=2.0.0
python-dotenv>=1.0.1
google-cloud-aiplatform>=1.69.0
google-cloud-storage>=2.18.0
neo4j>=5.21.0

# Heavy ML dependencies - install separately if needed
# Uncomment these lines if you need ML capabilities:
# torch>=2.0.0
# transformers>=4.30.0
# accelerate>=0.20.0
# sentencepiece>=0.1.99